connexité

+++ Arete.h
@@ -8,7 +8,7 @@
 class Arete
 {
     public:
-        Arete(grman::WidgetBox& sommet1,grman::WidgetBox& sommet2,int lsommet1,int lsommet2,int facteur);
+        Arete(grman::WidgetBox& sommet1,grman::WidgetBox& sommet2,int lsommet1,int lsommet2,float facteur);
         virtual ~Arete();
         void update(std::vector<Sommet*> Sommets);
         std::string save(std::vector<int> sommets_non_utilise);
